feat(lotto) print, map, util function들 모듈화
